Added a setupTile helper functions
this is so I can yank out the soon to be (HUGE) amounts of code setting
up the tiles and the setupGame() function is still able to be used.

NOTE: I have run out of time to setup the actual board, I may find time
sporatically throughout the day, otherwise it will be done tommorrow
morning.