Tile Creation Changes
TGiles are now created (distinctly) upon board constructor, same thing
will be done with sites and treasures, the game however will place these
elements.

I decided to do this becasue technically tyhey are distince pieces, so
their creation will never change with the board, however the location
may change.

The game should be handling those unique to a game, not the creation of
the actual pieces.

I will be branching to change the rest.